+++ sdk/cosmos/azure-cosmos/CHANGELOG.md
@@ -5,7 +5,7 @@
 #### Features Added
 * Added an API to retrieve diagnostics from the change feed processor context. - See [PR 41738](https://github.com/Azure/azure-sdk-for-java/pull/41738)
 * Added support to allow `queryChangeFeed` to complete when all changes available when the query starts have been fetched. - See [PR 42160](https://github.com/Azure/azure-sdk-for-java/pull/42160)
-* Added `cosmosFullTextPolicy` in `cosmosContainerProperties` and `cosmosFullTextIndexes` in `indexingPolicy` to support Full Text Search in CosmosDB - See [PR 42278](https://github.com/Azure/azure-sdk-for-java/pull/42278) 
+* Added `CosmosFullTextPolicy` in `CosmosContainerProperties` and `CosmosFullTextIndexes` in `IndexingPolicy` to support Full Text Search in Cosmos DB - See [PR 42278](https://github.com/Azure/azure-sdk-for-java/pull/42278)
 
 #### Breaking Changes